Delete some factories
